feat[docker-compose]: add script to create a full system
This will now spin up multiple instances of Node-Red so that one can be used for the MQTT client and two raspberry pi's to represent different sites. It also spins up an MQTT broker (ActiveMQ) and a database (InfluxDB). For Node-Red on the raspberry pi's the web interface port has changes to 4000 and 4001 for site one and two respectively.