Back end dev (#45)
* prot: add prototype version 1

* feat[docker-compose]: add script to create a full system

This will now spin up multiple instances of Node-Red so that one can be used for the MQTT client and two raspberry pi's to represent different sites. It also spins up an MQTT broker (ActiveMQ) and a database (InfluxDB). For Node-Red on the raspberry pi's the web interface port has changes to 4000 and 4001 for site one and two respectively.

* feat[influxdb]: changed port for working on linux

* feat[docker compose]: added the /data link and testing automatic flow insertion

* feat[Node_Red Flows]: add temp and flow_rate flows

* feat[docker]: folder structure update

* Create jsonFlow.json

feat [Node_Red_Flows]: Add JSON flow

* feat[Node_Red_Flows]: add end to end testing flow

* feat[Node_Red_Flows]: Update end to end testing flow

* feat[Node_Red_Flows]: Update end to end testing flow

* feat[Node_Red_Flows]: Update end to end testing flow. Add randomisation

* feat[Node_Red_Flows]: Add end to end v2

* feat[Node_Red_Flows]: Add UI flow

* Feat[Node Red Flows]: Add final flow

* Update[]:add test plan and delete docker-compose

Co-authored-by: Matt Dear <mrmdear@gmail.com>